+++ studio/src/app/components/editor/actions/deck/app-actions-deck/app-actions-deck.tsx
@@ -19,6 +19,7 @@ import {DemoAction} from '../../../../../utils/editor/demo-action';
 import {AnonymousService} from '../../../../../services/editor/anonymous/anonymous.service';
 import {OfflineService} from '../../../../../services/editor/offline/offline.service';
 import {RemoteService} from '../../../../../services/editor/remote/remote.service';
+import {PlaygroundAction} from '../../../../../utils/editor/playground-action';
 
 @Component({
   tag: 'app-actions-deck',
@@ -139,6 +140,8 @@ export class AppActionsDeck {
           await this.openPoll();
         } else if (detail.data.template === SlideTemplate.SPLIT && detail.data.attributes && detail.data.attributes.type === SlideSplitType.DEMO) {
           await this.openDemo();
+        } else if (detail.data.template === SlideTemplate.PLAYGROUND) {
+          await this.openPlayground();
         }
 
         if (detail.data.slide) {
@@ -169,6 +172,22 @@ export class AppActionsDeck {
 
     modal.onDidDismiss().then(async (detail: OverlayEventDetail) => {
       await this.addSlideYoutube(detail.data);
+
+      this.blockSlide.emit(false);
+    });
+
+    this.blockSlide.emit(true);
+
+    await modal.present();
+  }
+
+  private async openPlayground() {
+    const modal: HTMLIonModalElement = await modalController.create({
+      component: 'app-playground',
+    });
+
+    modal.onDidDismiss().then(async (detail: OverlayEventDetail) => {
+      await this.addSlidePlayground(detail.data);
 
       this.blockSlide.emit(false);
     });
@@ -253,6 +272,16 @@ export class AppActionsDeck {
 
       resolve();
     });
+  }
+
+  private async addSlidePlayground(playground: PlaygroundAction) {
+    if (!playground || !playground.src || playground.src === undefined || playground.src === '') {
+      return;
+    }
+
+    const slide: JSX.IntrinsicElements = await CreateSlidesUtils.createSlidePlayground(playground.src, playground.theme);
+
+    this.addSlide.emit(slide);
   }
 
   private addSlideDemo(demo: DemoAction): Promise<void> {