+++ studio/src/app/components/editor/actions/element/app-actions-element/app-actions-element.tsx
@@ -19,6 +19,7 @@ import {SlotUtils} from '../../../../../utils/editor/slot.utils';
 import {EditAction} from '../../../../../utils/editor/edit-action';
 import {MoreAction} from '../../../../../utils/editor/more-action';
 import {DemoAction} from '../../../../../utils/editor/demo-action';
+import {PlaygroundAction} from '../../../../../utils/editor/playground-action';
 
 import {BusyService} from '../../../../../services/editor/busy/busy.service';
 
@@ -630,13 +631,12 @@ export class AppActionsElement {
     await popover.present();
   }
 
-  private async openEditYoutubeSlide() {
-    if (this.slideNodeName !== 'deckgo-slide-youtube') {
-      return;
-    }
-
+  private async openEditModalSlide(
+    component: 'app-youtube' | 'app-demo' | 'app-playground',
+    onDismiss: (result: string | DemoAction | PlaygroundAction) => Promise<void>
+  ) {
     const modal: HTMLIonModalElement = await modalController.create({
-      component: 'app-youtube',
+      component,
       componentProps: {
         selectedElement: this.selectedElement,
       },
@@ -644,32 +644,7 @@ export class AppActionsElement {
 
     modal.onDidDismiss().then(async (detail: OverlayEventDetail) => {
       if (detail && detail.data && this.selectedElement) {
-        await this.updateYoutube(detail.data);
-      }
-
-      this.blockSlide.emit(false);
-    });
-
-    this.blockSlide.emit(true);
-
-    await modal.present();
-  }
-
-  private async openEditDemoSlide() {
-    if (!this.slideDemo) {
-      return;
-    }
-
-    const modal: HTMLIonModalElement = await modalController.create({
-      component: 'app-demo',
-      componentProps: {
-        selectedElement: this.selectedElement,
-      },
-    });
-
-    modal.onDidDismiss().then(async (detail: OverlayEventDetail) => {
-      if (detail && detail.data && this.selectedElement) {
-        await this.updateSlideDemo(detail.data);
+        await onDismiss(detail.data);
       }
 
       this.blockSlide.emit(false);
@@ -884,7 +859,7 @@ export class AppActionsElement {
     await this.imageHelper.deleteSlideAttributeImgSrc(this.selectedElement);
   }
 
-  private updateYoutube(youtubeUrl: string): Promise<void> {
+  private updateYoutube = (youtubeUrl: string): Promise<void> => {
     return new Promise<void>(async (resolve) => {
       if (!this.selectedElement || this.slideNodeName !== 'deckgo-slide-youtube') {
         resolve();
@@ -901,9 +876,33 @@ export class AppActionsElement {
 
       resolve();
     });
-  }
+  };
 
-  private async updateSlideDemo(demoAttr: DemoAction): Promise<void> {
+  private updatePlayground = async (playground: PlaygroundAction) => {
+    if (!this.selectedElement || this.slideNodeName !== 'deckgo-slide-playground') {
+      return;
+    }
+
+    if (!playground) {
+      return;
+    }
+
+    if (!playground.src || playground.src === undefined || playground.src === '') {
+      return;
+    }
+
+    this.selectedElement.setAttribute('src', playground.src);
+
+    if (playground.theme) {
+      this.selectedElement.setAttribute('theme', playground.theme);
+    } else {
+      this.selectedElement.removeAttribute('theme');
+    }
+
+    this.slideDidChange.emit(this.selectedElement);
+  };
+
+  private updateSlideDemo = async (demoAttr: DemoAction): Promise<void> => {
     if (!this.selectedElement || !this.slideDemo) {
       return;
     }
@@ -922,7 +921,7 @@ export class AppActionsElement {
     demo.setAttribute('mode', demoAttr.mode);
 
     this.slideDidChange.emit(this.selectedElement);
-  }
+  };
 
   private toggleList(destinationListType: SlotType.OL | SlotType.UL): Promise<void> {
     return new Promise<void>(async (resolve) => {
@@ -986,6 +985,7 @@ export class AppActionsElement {
       this.slideNodeName === 'deckgo-slide-chart' ||
       this.slideNodeName === 'deckgo-slide-poll' ||
       this.slideNodeName === 'deckgo-slide-youtube' ||
+      this.slideNodeName === 'deckgo-slide-playground' ||
       this.slideNodeName === 'deckgo-slide-author' ||
       this.slideDemo
     );
@@ -1099,9 +1099,11 @@ export class AppActionsElement {
           this.slideNodeName === 'deckgo-slide-poll'
             ? this.openEditPollSlide()
             : this.slideNodeName === 'deckgo-slide-youtube'
-            ? this.openEditYoutubeSlide()
+            ? this.openEditModalSlide('app-youtube', this.updateYoutube)
+            : this.slideNodeName === 'deckgo-slide-playground'
+            ? this.openEditModalSlide('app-playground', this.updatePlayground)
             : this.slideDemo
-            ? this.openEditDemoSlide()
+            ? this.openEditModalSlide('app-demo', this.updateSlideDemo)
             : this.openEditSlide()
         }
         color="primary"